Move unused memory utilities to helpers package
